Initial version of the OpenFOAM Cloud backend application
This application includes the entire logic for the OpenFOAM Cloud
application logic responsible for creating new OpenFOAM VMs and running
them on OpenStack. The application allows for customisation of the
underlying services used in the background, for example InfluxDB, snap
telemetry, S3 configuration, etc. There are no defaults for these
settings. Instead, an example local settings is provided for users to
see which options are available.

Basic documentation is provided on how to setup and run the application.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>