Decoupled saving of simulation and instance models from starting the experiment
Reorganization of start experiment flow. We first create DB
entries for simulation and instance models. Afterwards these entities are read
from the DB and used for spawning new VM instances that will run the simulation.
This way we will be able to use scheduled running of VM machines in the future.
Image and Server names are now created using UUID. This prevents nova from
returning servers with the same name as provided that were created
earlier/by someone else.

Change-Id: I81f1e4c5f4ab71e37c43db201cac35658965c490
Signed-off-by: Gasper Vrhovsek <gasper.vrhovsek@gmail.com>