Basic implementation of a scheduler daemon.
This commit introduces a basic implementation of a scheduler daemon. All
Simulation models are by default createt with 'PENDING' value in the status
field. This scheduler daemon queries these Simulation models each N seconds.
When it finds such simulation models in the DB it creates an osv image,
uploads it to glance and starts new servers on which the openFoam simulations
are calculated. Upon restarting the openfoam application the existing scheduler
(if any) gets killed and a new one is started.

Change-Id: I8bfdf39d4f54997637f7654272a19e7a6e4a3600
Signed-off-by: Gasper Vrhovsek <gasper.vrhovsek@gmail.com>