Implement use of NFS mounted directory for case and result location.
Allows for use of nfs mounted directory as a location for storing
simulation case files and also simulation results. Case files are
prepared for each instance with their respective modifications in
dedicated nfs mounted folders. After booting up OSv instances those
folders are mounted and used for simulation execution. Locations
of the results are saved in the database for each instance, so we
can access them later.

Change-Id: Iad553511cb74ef4018ec610d812e8e413ec81a0a
Signed-off-by: Gasper Vrhovsek <gasper.vrhovsek@gmail.com>