Launch simulation instnaces individually.
Instead of launching all instances linked to a certain simulation in a batch, we now
launch instances seperately, independent of the parent simulation. Each simulation then
finishes it's case and uploads the results to a dedicated folder on the NFS. We also enabled
retrying of failed simulations. Max retry count is configurable through local_settings.py

Change-Id: I318c111f7a3acd9edc6bc89061a25bca1eb7bc74
Signed-off-by: Gasper Vrhovsek <gasper.vrhovsek@gmail.com>